Fixed some quality of life stuff
Adds failure indicators, and better indicators that steps weren't run during early termination.

Adds the step name to the JSON structure, rather than relying on the index in the array.

Adds a load button to the start page, so we can load the same file without having to open the file dialog each time