Add variables to input files
Adds variable handling, which allows us to specify tests that depend on some data that will only be known at runtime. Adding this handling allows us to specify the data at runtime and dynamically replace the variable data in the step instructions. If the variable is not in the variable array for the test, but is in the instruction, it is ignored. Tests without variables display the same behaviour as before.